mcp_server/pyproject.toml
[project]
name = "business-intelligence-mcp"
version = "1.0.0"
description = "MCP сервер для бизнес-аналитики и рыночных данных"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
]
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"